Only try removing the metaDB file if it exists to avoid a warning

+++ t/URT/DataSource/Meta.pm
@@ -28,7 +28,7 @@ sub server {
 
 END {
     our $PATH;
-    unlink $PATH;
+    unlink $PATH if ($PATH);
 }
